add vault not found exception

+++ src/api/Records.js
@@ -1,6 +1,7 @@
 import axios from 'axios';
 import { notebookUrl } from './Notebooks';
 import { isTokenValid } from '../http';
+import VaultNotFoundException from '../CustomErrors/VaultNotFoundException';
 
 const url = '/records'
 
@@ -13,6 +14,8 @@ const Records = {
                 return this.getNotebookByName(vaultId)
                     .then(res => res.result.notebook)
                     .then(res => {
+                        if(!res)
+                            throw new VaultNotFoundException()
                         return axios.post(this.baseUrl + notebookUrl + '/' + res.ID + url, {
                             fields: recordFields,
                             notebookID: res.ID
@@ -37,6 +40,8 @@ const Records = {
                 return this.getNotebookByName(vaultId)
                     .then(res => res.result.notebook)
                     .then(res => {
+                        if(!res)
+                            throw new VaultNotFoundException()
                         return axios.get(this.baseUrl + notebookUrl + '/' + res.ID + url + '/' + recordId,
                             {
                                 headers: this.defaultHeaders
@@ -57,6 +62,8 @@ const Records = {
                 return this.getNotebookByName(vaultId)
                     .then(res => res.result.notebook)
                     .then(res => {
+                        if(!res)
+                            throw new VaultNotFoundException()
                         return axios.put(this.baseUrl + notebookUrl + '/' + res.ID + url + '/' + recordId, {
                             fields: recordFields,
                             notebookID: res.ID
@@ -78,6 +85,8 @@ const Records = {
                 return this.getNotebookByName(vaultId)
                     .then(res => res.result.notebook)
                     .then(res => {
+                        if(!res)
+                            throw new VaultNotFoundException()
                         return axios.delete(this.baseUrl + notebookUrl + '/' + res.ID + url + '/' + recordId
                             , {
                                 headers: this.defaultHeaders
@@ -99,6 +108,8 @@ const Records = {
                 return this.getNotebookByName(vaultId)
                     .then(res => res.result.notebook)
                     .then(res => {
+                        if(!res)
+                            throw new VaultNotFoundException()
                         return axios.post(this.baseUrl + notebookUrl + '/' + res.ID + url, {
                             records: recordFields,
                             notebookID: res.ID
